# Speedup nixos package count