@nonetrix 2024-05-01
Great work, will be switching over to Fastfetch because of this just hoping it gets merged into nixpkgs unstable soon. 30ms does still seem a little slow compared to Arch for example but MUCH better than it was before, but is the any more room to reduce it further? I wish I could just cache the result and only update it when the system is rebuilt to be honest